SourceCode/Final Code/cardio-disease-network/backend-bayesian/requirements.txt
fastapi
uvicorn
pandas
pgmpy
networkx
pydantic
google-genai
python-dotenv